addSortAndLimit functions
